metricas-onvio/backend/requirements.txt
fastapi
uvicorn
python-multipart
pandas
openpyxl
google-cloud-bigquery
python-jose[cryptography]
passlib[bcrypt]
pydantic
pydantic-settings
email-validator
requests
httpx
db-dtypes
